# api/src/main/resources/api-definition.yaml
openapi: 3.0.3
info:
  contact:
    name: Sandro Andreasik
  description: Sprint Planning Poker for Scrum teams
  title: EFI Poker API
  version: 1.0.0
servers:
- description: API base path
  url: /api/v1
...
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
          description: User profile
      security:
      - bearerAuth: []
      summary: Get current authenticated user
      tags:
      - auth
...
  /auth/oauth2/callback:
    get:
      description: |
        Callback endpoint for Auth0 after successful authentication. Exchanges authorization code for tokens, finds or creates the user, issues an application JWT, and redirects back to the frontend.
      operationId: oauth2Callback
      parameters:
      - description: Authorization code returned by Auth0
        in: query
        name: code
        required: false
        schema:
          type: string
      - description: State parameter for CSRF protection validation
        in: query
        name: state
        required: false
        schema:
          type: string
      - description: Error code if the user denies authentication
        in: query
        name: error
        required: false
        schema:
          type: string
      responses:
        "302":
          description: Redirect to frontend with JWT token in query parameter
          headers:
            Location:
              description: Frontend URL with JWT token (e.g. /auth/callback?token=eyJ...)
              schema:
                format: uri
                type: string
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad request (missing code or state mismatch)
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Authentication failed (invalid or expired code)
      security: []
      summary: Auth0 OAuth2 callback
      tags:
      - auth
...
  /admin/users:
    get:
      description: Returns a paginated list of all users. Requires ADMIN role.
      operationId: listUsers
      parameters:
      - description: Page number (0-based)
        in: query
        name: page
        required: false
        schema:
          default: 0
          type: integer
      - description: Page size
        in: query
        name: size
        required: false
        schema:
          default: 20
          type: integer
      - description: Search by username or email
        in: query
        name: search
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedUsersResponse"
          description: List of users
      security:
      - bearerAuth: []
      summary: List all users
      tags:
      - admin
    post:
      description: Admin creates a new local user account. Requires ADMIN role.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateUserRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
          description: User created
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Username already taken
      security:
      - bearerAuth: []
      summary: Create a new user
      tags:
      - admin
  /admin/users/{id}:
    delete:
      description: Delete a user account. Cannot delete yourself. Requires ADMIN role.
      operationId: deleteUser
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "204":
          description: User deleted
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: User not found
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Cannot delete yourself
      security:
      - bearerAuth: []
      summary: Delete user
      tags:
      - admin
    get:
      description: Returns details of a specific user. Requires ADMIN role.
      operationId: getUser
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
          description: User details
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: User not found
      security:
      - bearerAuth: []
      summary: Get user details
      tags:
      - admin
    patch:
      description: Update user role or email. Requires ADMIN role.
      operationId: updateUser
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
          description: User updated
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: User not found
      security:
      - bearerAuth: []
      summary: Update user
      tags:
      - admin
components:
...
  schemas:
...
    ProblemDetail:
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      type: object
...
    UserResponse:
      example:
        role: ADMIN
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        email: email
        username: username
      properties:
        id:
          format: uuid
          type: string
        username:
          type: string
        email:
          type: string
        role:
          enum:
          - ADMIN
          - USER
          type: string
      type: object
    PagedUsersResponse:
      example:
        size: 5
        totalPages: 6
        page: 1
        content:
        - authProvider: authProvider
          createdAt: 2000-01-23T04:56:07.000+00:00
          role: ADMIN
          lastLoginAt: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          email: email
          username: username
        - authProvider: authProvider
          createdAt: 2000-01-23T04:56:07.000+00:00
          role: ADMIN
          lastLoginAt: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          email: email
          username: username
        totalElements: 0
      properties:
        content:
          items:
            $ref: "#/components/schemas/AdminUserResponse"
          type: array
        totalElements:
          format: int64
          type: integer
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer
      type: object
    AdminCreateUserRequest:
      example:
        password: password
        role: ADMIN
        email: email
        username: username
      properties:
        username:
          maxLength: 100
          minLength: 3
          type: string
        password:
          maxLength: 128
          minLength: 8
          type: string
        email:
          format: email
          type: string
        role:
          enum:
          - ADMIN
          - USER
          type: string
      required:
      - password
      - role
      - username
      type: object
    AdminUserResponse:
      example:
        authProvider: authProvider
        createdAt: 2000-01-23T04:56:07.000+00:00
        role: ADMIN
        lastLoginAt: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        email: email
        username: username
      properties:
        id:
          format: uuid
          type: string
        username:
          type: string
        email:
          type: string
        role:
          enum:
          - ADMIN
          - USER
          type: string
        authProvider:
          type: string
        createdAt:
          format: date-time
          type: string
        lastLoginAt:
          format: date-time
          type: string
      type: object
    AdminUpdateUserRequest:
      example:
        role: ADMIN
        email: email
      properties:
        email:
          format: email
          type: string
        role:
          enum:
          - ADMIN
          - USER
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http